feat/refactor: defined values of return in trait Audio and impl for Wav Format

+++ src/audios/mod.rs
@@ -13,7 +13,7 @@ pub trait Audio {
 
     fn sample_rate(&self) -> u32;
 
-    fn bit_depth(&self) -> u8;
+    fn bit_depth(&self) -> u16;
 
-    fn channels(&self) -> u8;
+    fn channels(&self) -> u16;
 }
